Move landing page files into Landing Page sub folder
