+++ src/components/LandingPage/header-bar.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect, } from 'react-redux';
 import { slide as Menu, } from 'react-burger-menu';
-import { clearAuth, } from '../actions/auth';
+import { clearAuth, } from '../../actions/auth';
 import { showLoginForm, } from '../actions/users';
 import { clearAuthToken, } from '../local-storage';
 import { hideLoginForm, } from '../actions/users';
