+++ src/components/LandingPage/landing-page.css
@@ -74,11 +74,7 @@ margin-bottom: 300px;
   background: #00d2ff; /* fallback for old browsers */
   background: -webkit-linear-gradient(to left, #00d2ff, #3a7bd5); /* Chrome 10-25, Safari 5.1-6 */
   background: linear-gradient(to left, #00d2ff, #3a7bd5); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
-color: white;
-font-family: 'Lato';
-line-height: 2;
+  color: white;
+  font-family: 'Lato';
+  line-height: 2;
 }
-
-/* .bm-burger-button {
-background-color: black;
-} */
