+++ src/components/LandingPage/modal.css
@@ -8,6 +8,5 @@
   width: 100%;
   height: 100%;
   overflow: auto;
-  background-color: rgb(0,0,0);
   background-color: rgba(0,0,0,0.7); 
 }
