+++ src/components/LandingPage/registration-page.js
