+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router, } from 'react-router-dom';
 import { Provider, } from 'react-redux';
-import App from './components/app';
+import App from './components/LandingPage/app';
 import store from './store';
 import './index.css';
