+++ src/reducers/landingpage.js
@@ -8,7 +8,6 @@ const initialState = {showLoginForm: false,};
 export function landingPageReducer(state = initialState, action) {
   switch (action.type) {
   case SHOW_LOGIN_FORM:
-    console.log(action.type);
     return {
       ...state,
       showLoginForm: true,
